backend/requirements.txt
beautifulsoup4==4.13.3
boto3==1.34.145
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dj-rest-auth==4.0.1
django-allauth==0.56.1
django-cacheops==7.0.2
django-crontab==0.7.1
django-dbbackup==4.1.0
django-debug-toolbar==4.3.0
django-filter==23.3
django-redis==5.4.0
django-storages==1.14.4
google-api-python-client==2.139.0
gunicorn==22.0.0
Markdown==3.7
social-auth-app-django==5.3.0
Pillow==9.2.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1